- Use req->rq_export->exp_obd, not req->rq_obd, to find the MDS for a request,   since the latter is now the MDT. - transnos are 64 bits. - fix some wacky whitespace in class_obd.c

+++ lustre/ptlrpc/recover.c
@@ -159,7 +159,7 @@ int ptlrpc_replay(struct obd_import *imp)
         struct list_head *tmp, *pos;
         struct ptlrpc_request *req;
         struct ptlrpc_connection *conn = imp->imp_connection;
-        int committed = imp->imp_peer_committed_transno;
+        __u64 committed = imp->imp_peer_committed_transno;
         ENTRY;
 
         spin_lock(&imp->imp_lock);
